Only add `quantity` or `units` keys in steps if they are defined
Prior to this change, if a quantity or a unit is undefined for an
ingredient, cookware, or timer, then the resulting parsed direction item
would still include the `quantity` or `unit` key despite the value being
undefined.

For example, given the following ingredient:

    @cauliflower{1}

The resulting parsed direction item would be:

    {
      "type": "ingredient",
      "name": "cauliflower",
      "quantity": 1,
      "units": undefined
    }

The problem with this is that it makes it trickier to convert the
resulting steps structure to JSON using `JSON.stringify()`, as undefined
is not a native JSON value:

    Reason: `undefined` cannot be serialized as JSON. Please use `null`
    or omit this value.

This commits causes the keys to be omitted from the structure entirely
if their values are undefined:

    {
      "type": "ingredient",
      "name": "cauliflower",
      "quantity": 1
    }

The `quantity` and `units` keys in the `StepIngredient`, `StepCookware`
and `StepTimer` types are already marked as optional, so omitting the
keys entirely should not cause any issues.