Bring default ingredient/cookware quantities in line with spec
Following on from the discussion in cooklang/cooklang-ts#2, this commit
sets the default cookware quantity to `1`, and the default ingredient
quantity to `"some"` - rather than their previous defaults of
`undefined`.

This also introduces a new Parser option, `defaultCookwareAmount`, which
allows customisation of the default cookware quantity - thus allowing
users to override the default quantity of `1` if they wish.

Avoiding `"quantity": undefined` is important to allow the recipe
structure to be neatly serialised as JSON, which doesn't support the
concept of `undefined` values.