Bring default ingredient/timer units in line with spec
Following on from the discussion in cooklang/cooklang-ts#2, this commit
sets the default units for both ingredients and timers to `""` - rather
than their previous defaults of `undefined`.

Avoiding `"units": undefined` is important to allow the recipe structure
to be neatly serialised as JSON, which doesn't support the concept of
`undefined` values.